+++ app/src/store/storageSlice.ts
@@ -1,17 +1,5 @@
 import { pick } from 'lodash'
 import { StateCreator } from 'zustand'
-import { LensDataState } from './lensSlice'
-import { TableState } from './tableSlice'
-
-export type LocalStorageData = {
-    state: Pick<TableState & LensDataState, 'lenses' | 'units' | 'distance' | 'order' | 'orderBy'>
-    version: number
-}
-
-export interface StorageState {
-    extractForLocalStorage: () => LocalStorageData
-    applyFromLocalStorage: (partialState: LocalStorageData) => void
-}
 
 export const createStorageSlice: StateCreator<TableState & LensDataState & StorageState, [], [], StorageState> = (
     _set,
